add sold_units processing

+++ db/migrate/20161012134329_update_received_inventories.rb
@@ -4,6 +4,6 @@ class UpdateReceivedInventories < ActiveRecord::Migration[5.0]
     add_column :received_inventories, :sold_units, :integer
     add_column :received_inventories, :cost_sold, :decimal, precision: 10, scale: 2 # revenue
     add_column :received_inventories, :remain_units, :integer
-    add_column :received_inventories, :cost_remain, :integer, precision: 10, scale: 2
+    add_column :received_inventories, :cost_remain, :decimal, precision: 10, scale: 2
   end
 end
